% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb2google.R
\name{bb2google}
\alias{bb2google}
\title{Convert a bounding box into the data needed by Google Maps to create a basemap.}
\usage{
bb2google(bb)
}
\arguments{
\item{bb}{A numeric vector with four elements representing a lat,lon bounding box}
}
\value{
A pair of numeric values for the map center and an integer for the zoom level
}
\description{
Bounding boxes are created from a set of geographic coordinates using the
\code{make_bbox} function (from the ggmap package).

The bounding box may not be square. The \code{bb_square} function (from mapfunctions)
fixes that.

This function takes a bounding box and returns the center coordinates and
the zoom level so you can create a basemap with the \code{make_google_basemap} function
(from mapfunctions).
}
\examples{
x <- c(left=150.1,bottom=21.3,right=135.1,top=35.3)
g_info <- bb2google(bb=x)
map_center <- g_info$center
zoom_level <- g_info$zoom
}
