% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_cuts.R
\name{site_cuts}
\alias{site_cuts}
\title{Divides a set of quantitative values into qualitative segments based on a particular strategy.}
\usage{
site_cuts(quant_var, cuttype, cuts = NA)
}
\arguments{
\item{quant_var}{The table column used as the basis of the segmentation}

\item{cuttype}{The strategy to be used}

\item{cuts}{The list of breakpoints (cutlist) or number of cuts (breaks)}
}
\value{
List of index values (1 to nrows) for the segments
}
\description{
Quantitative values often need to be divided into categories. How the
breaking into segments occurs depends on the goal of the visualization.

This function provide for seven alternative strategies.
\itemize{
\item topbottom: divide the range in half and assign row.
\item fiftyfifty: half the rows into the first group and the top in the second.
\item quartile3: top and bottom quartiles with the two middle as the other group.
\item quartile4: each quartile a separate group.
\item statistics3: below and above one standard deviation plus the middle.
\item statistics4: like statistics3 but separate the middle at the mean.
\item cutlist: use cut list of values for cut locations (min and max added automatically).
\item breaks: use the value of the cuts for the number of segments
}

This function doesn't use the entire datatable. Instead, just a single column
of values is passed to the function (quant_var) along with a strategy for
determining the cut locations (i.e., one of the seven listed above). If the breaks
strategy is specified, the cuts parameter will be a
list of cut locations at which the data values will be divided.

The returned value is a list with one row for each of the data value rows. This
list has an index number representing the category for the row value.
}
