% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_elevations.R
\name{site_elevation}
\alias{site_elevation}
\title{Return the elevations for a datatable with geographic coordinates.}
\usage{
site_elevation(datatable, APIkey)
}
\arguments{
\item{datatable}{The standard datatable with columns for lat, lon}

\item{APIkey}{A Google Maps API key}
}
\value{
datatable with elevations in a column named 'elevation'
}
\description{
This uses Google Maps to get the elevations of locations in a datatable.

The process involves an API call to Google Maps. As a result, you need
to have a valid Google Maps Key. This key is sent to Google Maps each time
you ask for an elevation. The explicit use of the key in this function
makes it a bit different than the other sitemap functions that use the Google
Maps server.
}
