% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_fans.R
\name{site_fans}
\alias{site_fans}
\title{Create map symbols that show a direction and field of view.}
\usage{
site_fans(datatable, styles = column)
}
\arguments{
\item{datatable}{The standard datatable with columns for lat, lon,
fan_direction and fan_field_width}

\item{styles}{The standard complement of column values}
}
\value{
Data supplied to ggmap that places a layer with symbols on the basemap
}
\description{
Photo locations on a map may be more useful if the direction (bearing)
of the photo and its field of view are part of the symbol marking the
place where the photo was taken. This is called a 'fan.'

The minimal data, beyond the normal data point info (i.e., lon, lat),
is the fan_direction and fan_field_width (i.e., the field of view).
Each of these specifications is either in
a column in a data table or, if the value is constant, a column
style value (e.g., column$field_width <- 48).

There are a number of styles that go with the site_fans function. The
default values are loaded with site_styles.
}
