% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_google_hides.R
\name{site_google_hides}
\alias{site_google_hides}
\title{Load a set of parameter values used to hide layers on Google Maps basemaps.}
\usage{
site_google_hides()
}
\value{
A set of values which can be used to hide map layers
}
\description{
Google generated basemaps often contain layers of information, such as
road names, points of interest, and landscape features. These layers can
generally be hidden.

This function brings in code that tells the Google Map server to suppress the
various layers.

Load this data early in the code. Note that this does not load with \code{set_param()}.

Consult the vignettes to see how to use these control values.
}
\examples{
hide <- site_google_hides()

}
