% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_legend.R
\name{site_legend}
\alias{site_legend}
\title{Create a legend to explain the symbolism used with the map points.}
\usage{
site_legend(h_table, p_table, styles = column)
}
\arguments{
\item{h_table}{A data frame with row number, text, and styles for each header row}

\item{p_table}{A data frame with row number, point characteristics, and meaning}

\item{styles}{A dataframe with style values for each of the map characteristics}
}
\value{
Plot with the map legend
}
\description{
There are many ways map points can be colored, sized, and outlined.
These attributes provide visual cues to information. The link between
the point attributes and the meaning is made in the map legend.

The legend is constructed from a rows of information. Some rows have
descriptive text, such as identifying a category of points, such as
points of different sizes representing population sizes. The other rows
have a point symbol and a short text statement that describes the
attribute associated with the specific symbolism. For size, there may
be a set of different size points with the terms small, medium and
large.

The function uses two tables as primary input. The header table (h_table) has
the text lines. The point table (p_table) has the symbol information. Each
table has line numbers so that the header text and points are properly
merged to make the legend.

Values that are constant in a table can be entered (or defaulted) using
the column values.

Note that this function does not handle legend information for labels.
}
