% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_names.R
\name{site_names}
\alias{site_names}
\title{Add names (i.e., words) to a basemap layer.}
\usage{
site_names(datatable, styles = column)
}
\arguments{
\item{datatable}{The table with the text, location and (optionally) style data}

\item{styles}{A dataframe with style values for each of the map characteristics}
}
\value{
Data supplied to ggmap that places a layer on the basemap
}
\description{
Basemaps of come with new, if any, names to mark important locations. It is
useful to add names so that a map user is properly oriented.
At a minimum, each name has a text value (what will appear on the map) and
a pair of coordinate values (lon,lat) showing the location of the text.

Each text value can be customized in many ways, including size, color,
font face, and font type.

Text customization can be done with name styles (e.g., column$name_text_color)
when all text is to have the same value or (more commonly) a table column
(e.g., name_text_color) is used to allow the text for each name to have its own
style characteristics.

A dataframe (called column) stores a set of default style values. Load this
set by using the function site_styles().
}
