% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_photos.R
\name{site_photos}
\alias{site_photos}
\title{Reads EXIF data to get coordinates and other info from all the JPG files in a folder.}
\usage{
site_photos(folder)
}
\arguments{
\item{folder}{The name of the folder with a set of JPG images}
}
\value{
table_info with file number, name, description, lat, lon, direction,
view width, distance, date, time
}
\description{
You can get most digital phone cameras to record location information in the
EXIF data that is part of a JPG image. This function extract that data.

There are several useful fields that are retrieved along with the geographic
coordinates. These include the contents of the Description field, the date,
and the time of the photo.

A reference number (the file order in the folder) and the file name are also
returned.
}
