% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_points.R
\name{site_points}
\alias{site_points}
\title{Add points to a basemap layer.}
\usage{
site_points(datatable, styles = column)
}
\arguments{
\item{datatable}{The table with the text, location and (optionally) style data}

\item{styles}{A dataframe with style values for each of the map characteristics}
}
\value{
Data supplied to ggmap that places a layer on the basemap
}
\description{
It is common to mark important locations with points.

Each point can be customized in many ways, including size, color,
surrounding line, and more.

Point customization can be done with point parameters (e.g., column$point_color)
when all the points are to have the same style or (more commonly) by using
a table column (e.g., point_color) to allow the text for each name to have its own style
characteristic.

A dataframe (called column) stores a set of default values. Load this
set by using the function site_styles().
}
