% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_scale.R
\name{site_scale}
\alias{site_scale}
\title{Place a scale bar drawn as a basemap overlay to show the dimensions of the map.}
\usage{
site_scale(basemap, styles = column)
}
\arguments{
\item{basemap}{A basemap generated by ggmap}

\item{styles}{A data frame of default style values}
}
\value{
An annotated scale bar (line and text)
}
\description{
The size of a map is an important feature that is included on many maps.
This function automates the general process of placing a scale bar and
an associated text statement with the length and measurement units for
the scale bar.

The length of the scale bar is determined automatically. The general width
is controlled with a scale width parameter. The actual width of the scale
bar is determined by rounding the value of the bar length with a primary
greatest value algorithm.

The user can choose whether the dimensions should be metric (m, km) or
English (ft, mi).

As with the length of the scale bar, the text for the dimension is rounded
to the primary greatest value. In addition, the scaling shifts from smaller
units (m, ft) to larger units (km, mi) as the length of the bar increases.

A number of other display features are available using param values. These
include the text size and color, as well as line thickness and color.

The location of the scale can be chosen from the four alternatives:
ll, lr, ul, and ur.

Most of the style values will be defaulted using the param values.
}
