% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_squaremap.R
\name{site_squaremap}
\alias{site_squaremap}
\title{Convert a basemap bounding box into one that is square.}
\usage{
site_squaremap(bb)
}
\arguments{
\item{bb}{A numeric vector with four elements representing a lat,lon bounding box}
}
\value{
A numeric vector with bounding box values adjusted to make a square bounding box
}
\description{
Bounding boxes are created from the distribution of coordinate values. It is
not likely that the resulting bounding box will be square. This might be OK
for Stamen basemaps. But Google basemaps are always square and you'll need
to make the bounding box square in order to calculate the map center and
zoom level.
}
\examples{
x <- c(left=150.1,bottom=21.3,right=135.2,top=30.5)
square <- site_squaremap(bb=x)
}
