% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_stamen_basemap.R
\name{site_stamen_basemap}
\alias{site_stamen_basemap}
\title{Create a custom basemap using Stamen Maps as the server.}
\usage{
site_stamen_basemap(datatable, styles = column)
}
\arguments{
\item{datatable}{The table with the text, location and (optionally) style data}

\item{styles}{A dataframe with style values for some map characteristics}
}
\value{
A basemap that is scaled to hold the data points
}
\description{
A basemap is a key element in making a map. The Stamen Map server provides
so useful designs, particularly for maps that show country-size areas or
smaller.
}
