% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_styles.R
\name{site_styles}
\alias{site_styles}
\title{Set all the default style values used for basemaps and annotation.}
\usage{
site_styles()
}
\value{
A set of values which are used as defaults
}
\description{
Load these data early in the code. Once these are set, you can override any
columneter with a separate line of code, such as:

\code{column$point_color <- "yellow"}

Note that the names of the styles (e.g., \code{label_text_size}, \code{point_color}) are the names
of columns in the data table if you are specifying separate values for each
style in the table.

Consult the vignettes for information on the use of these styles
and the alternative values they may take.
}
\examples{
column <- site_styles()

}
